Scheduler moving location permission logic to AppointmentScreen
